requirements.txt
customtkinter
pyautogui
opencv-python
pillow
mss
fire
psutil
pygame
pytesseract
